define and document elsize
Fixes #36553
Fixes #37438
