+++ base/abstractarray.jl
@@ -149,7 +149,37 @@ valtype(A::Type{<:AbstractArray}) = eltype(A)
 prevind(::AbstractArray, i::Integer) = Int(i)-1
 nextind(::AbstractArray, i::Integer) = Int(i)+1
 
+
+"""
+    eltype(type)
+
+Determine the type of the elements generated by iterating a collection of the given `type`.
+For dictionary types, this will be a `Pair{KeyType,ValType}`. The definition
+`eltype(x) = eltype(typeof(x))` is provided for convenience so that instances can be passed
+instead of types. However the form that accepts a type argument should be defined for new
+types.
+
+# Examples
+```jldoctest
+julia> eltype(fill(1f0, (2,2)))
+Float32
+
+julia> eltype(fill(0x1, (2,2)))
+UInt8
+```
+"""
+eltype(::Type) = Any
+eltype(::Type{Bottom}) = throw(ArgumentError("Union{} does not have elements"))
+eltype(x) = eltype(typeof(x))
 eltype(::Type{<:AbstractArray{E}}) where {E} = @isdefined(E) ? E : Any
+
+"""
+    elsize(type)
+
+Compute the memory stride in bytes between consecutive elements of `eltype`
+stored inside the given `type`, if the array elements are stored densely with a
+uniform linear stride.
+"""
 elsize(A::AbstractArray) = elsize(typeof(A))
 
 """
