+++ base/array.jl
@@ -102,28 +102,6 @@ const DenseVecOrMat{T} = Union{DenseVector{T}, DenseMatrix{T}}
 
 ## Basic functions ##
 
-"""
-    eltype(type)
-
-Determine the type of the elements generated by iterating a collection of the given `type`.
-For dictionary types, this will be a `Pair{KeyType,ValType}`. The definition
-`eltype(x) = eltype(typeof(x))` is provided for convenience so that instances can be passed
-instead of types. However the form that accepts a type argument should be defined for new
-types.
-
-# Examples
-```jldoctest
-julia> eltype(fill(1f0, (2,2)))
-Float32
-
-julia> eltype(fill(0x1, (2,2)))
-UInt8
-```
-"""
-eltype(::Type) = Any
-eltype(::Type{Bottom}) = throw(ArgumentError("Union{} does not have elements"))
-eltype(x) = eltype(typeof(x))
-
 import Core: arraysize, arrayset, arrayref, const_arrayref
 
 vect() = Vector{Any}()
